feat: Sentiment Analyzer (#18)
Adds a sentiment analyzer using BERT to observe sentiment in selected channels (positive/neutral/negative) and sends a warning message in a separate channel in case sentiment is overwhelmingly negative.

The idea behind this being that this could be an additional tool to assist moderators in their task by pre-emptively warning them if sentiment in a channel turns negative.

This approach is not perfect and will have both false positives and negatives. It should just be considered an additional tool.